fix: 발송이력 권한 조회 오류 수정
- MyBatis OGNL에서 contains() Character 변환 오류 수정 (indexOf 사용)
- getUserLevel에서 JWT seq로 사용자 조회하도록 수정

+++ src/main/java/kr/wisead/mapper/primary/UserMapper.java
@@ -11,7 +11,7 @@ import org.apache.ibatis.annotations.Param;
 public interface UserMapper {
 
   /** 회원 조회 (by SEQ) */
-  Optional<User> findBySeq(@Param("seq") Long seq);
+  Optional<User> findBySeq(@Param("seq") Integer seq);
 
   /** 회원 조회 (by USER_ID) */
   Optional<User> findByUserId(@Param("userId") String userId);
